Added support for allowEditID in addition to regular calendarID. Not finished yet. Frontend todo: Finish having every http that changes a calendar to also send allowEditID.   Backend todo CREATE HELPER FUNCTIONS calendarExists (check if calendar exists), authorize (authorize calendarID with allowEditID).

+++ backend/app.js
@@ -16,7 +16,7 @@ const express = require('express');
 const helmet = require("helmet");
 const cors = require('cors')
 const corsOptions = {
-    origin: process.env.ALLOWED_CORS_ORIGIN,
+    origin: "*",
     credentials: true
   }
 const bodyParser = require('body-parser');
@@ -29,13 +29,15 @@ app.use(bodyParser.urlencoded({extended: true}));
 app.use(bodyParser.json());
 
 //API
+//TODO CREATE HELPER FUNCTIONS calendarExists (check if calendar exists), authorize (authorize calendarID with allowEditID).
 app.post('/api/newCalendar', (req, res, next) => {
     let calendarID = createID();
+    let allowEditID = createID();
     pool
-        .query('INSERT INTO calendar(calendarID, dateTimeCreated) VALUES($1, LOCALTIMESTAMP)', [calendarID])
+        .query('INSERT INTO calendar(calendarID, allowEditID, dateTimeCreated) VALUES($1, $2, LOCALTIMESTAMP)', [calendarID, allowEditID])
         .then(pgresult => {
             //If successful send the new calendar id
-            res.status(201).send(calendarID);
+            res.status(201).send(JSON.stringify({calendarID:calendarID, allowEditID:allowEditID}));
         })
         .catch(e => {console.error(e.stack);
             res.status(400).end();}
@@ -44,12 +46,26 @@ app.post('/api/newCalendar', (req, res, next) => {
 
 app.delete('/api/deleteCalendar', (req, res, next) => {
     let calendarID = req.query.calendarID;
+    let allowEditID = req.body.allowEditID;
+    //Checks if calendarID and allowEditID match up. If not, send forbidden.
     pool
-        .query('DELETE FROM calendar WHERE calendarID = $1 RETURNING *', [calendarID])
+        .query('SELECT COUNT(*) FROM calendar WHERE calendarID = $1 AND allowEditID = $2', [calendarID, allowEditID])
         .then(pgresult => {
-            //If successful send 200, if not found send 404
-            pgresult.rows[0] ? res.status(200).end() : res.status(404).end()
-            res.status(200).end();
+            if (Number(pgresult.rows[0].count) === 0) {
+                res.status(403).end();
+            }
+            else {
+                pool
+                .query('DELETE FROM calendar WHERE calendarID = $1 RETURNING *', [calendarID])
+                .then(pgresult => {
+                    //If successful send 200, if not found send 400
+                    pgresult.rows[0] ? res.status(200).end() : res.status(404).end()
+                    res.status(200).end();
+                })
+                .catch(e => {console.error(e.stack);
+                    res.status(400).end();}
+                )
+            }
         })
         .catch(e => {console.error(e.stack);
             res.status(400).end();}
@@ -73,8 +89,9 @@ app.post('/api/newEvent', (req, res, next) => {
     let calendarID = req.body.calendarID;
     let eventDescription = req.body.eventDescription;
     let starttime = req.body.starttime;
-    let length = req.body.length;
+    let length = 100;
     let eventID = createID();
+    let allowEditID = req.body.allowEditID;
     //temporary value for starttime
     starttime = Date.now();
 
@@ -103,6 +120,7 @@ app.get('/api/getEvent', (req, res, next) => {
 
 app.delete('/api/deleteEvent', (req, res, next) => {
     let eventID = req.query.eventID;
+    let allowEditID = req.body.allowEditID;
     pool
         .query('DELETE FROM event WHERE eventID = $1 RETURNING *', [eventID])
         .then(pgresult => {
@@ -120,7 +138,6 @@ app.get('/api/loadEventsFromCalendar', (req, res, next) => {
     pool
         .query('SELECT eventID FROM event WHERE calendarID = $1', [calendarID])
         .then(pgresult => {
-            console.log(pgresult.rows);
             res.send(pgresult.rows);
         })
         .catch(e => {console.error(e.stack);
